Move dummy data to a diff folder
